Sequelize setup finished
